+++ src/Layout/MainLayout/MainLayout.jsx
@@ -1,11 +1,9 @@
 import { Outlet } from "react-router-dom";
-import Footer from "../../Footer";
 
 const MainLayout = () => {
   return (
     <div>
       <Outlet />
-      <Footer></Footer>
     </div>
   );
 };
