+++ src/Pages/Home/Home.jsx
@@ -1,7 +1,10 @@
+import Footer from "../../components/Home/Footer/Footer";
+
 const Home = () => {
   return (
     <div>
       <h1 className="text-5xl text-center">Home</h1>
+      <Footer />
     </div>
   );
 };
